Setup block and filesystem layer
This attempts to probe for the W25Q128JV flash chips on the v0.3 board,
and failing to find them, falls back to looking for the microSD card on
the v0.2 board. It panics if it can find neither, and otherwise attempts
to use whatever storage it does find as the underlying block layer for a
littlefs filesystem, which gets mounted at "/data".