Introduce webrtc_manager
For now, this is mostly just moving code around, but the ultimate goal
is that the webrtc_manager will be responsible for maintaining the
WebRTC connection at all times (even when we're not actively playing it
back). That includes maintaining the target buffer size and consuming
packets when they aren't being sent to the mixer.