Rework ADC task
Always start/stop the ADC from adc_task, instead of whoever is
subscribing/unsubscribing. (This avoids a bug where the ADC crashes if
it's stopped from a different task than it's started from).
Additionally, add a watchdog that fires if we've received no new ADC
data for over 1.5 seconds, to try and catch whatever ADC lockup we've
been experiencing.